Apply ContextualAnalysis on python VEnv
